src/requirements.txt
boto3
aws-msk-iam-sasl-signer-python
confluent_kafka
python-dotenv